Kernel patch to add rootdelay feature
This patch adds the 'rootdelay' option to the kernel
command-line boot options. The feature was backported
from the 2.6 kernel series. This allows for mounting
root filesystems which are located on devices whose
drivers are slow to load, such as USB mass-storage devices.

Example: 'rootdelay=10' tells the kernel tp wait 10
seconds before trying to mount the rootfs device.