IDE: fix panic during probe with negative IRQ
I've attached the ide-probe.c patch.  It disallows not only negative
IRQs, but also IRQ 0 - but the original code did the same (and in other
places as well).  BTW, 2.6 has the same fishy code in ide-probe.c, so
maybe the problem is actually elsewhere (that is, maybe -1 is not a
valid value for that field, and maybe that is fixed in 2.6).