x86 SMP: don't report error on uniprocessor machines
Booting i386 on an SMP system with only one processor reports
the following error :

   Error: only one processor found.

This has always caused unnecessary worries to admins, and is
getting more frequent nowadays on virtual machines. The solution
simply consists in removing the cpucount test, which does not
exist anymore in 2.6 BTW. A boot on a quad-core "downcored"
to 1 core now correctly reports :

   Total of 1 processors activated (6003.09 BogoMIPS).

Ditto for x86_64.

Signed-off-by: Willy Tarreau <w@1wt.eu>