wdt: add support for AMD Geode GX/LX watchdog
This is a backport of AMD's Geode watchdog driver from 2.6 to 2.4.
This driver requires the MFGPT infrastructure, which itself is
enabled when CONFIG_MGEODE_LX is set. Loading this driver will
cause the MFGPT code to initialize the timers. If the system hangs
because of a broken BIOS, it is recommended to pass "mfgptfix" to
the kernel command line to first reset the registers.

Signed-off-by: Willy Tarreau <w@1wt.eu>