i386: add support for AMD Geode MFGPT timers
This is a second backport of the 2.6 Geode MFGPT code to 2.4. It
enables use of the Geode on-chip MFGPT timers which may be used
by a watchdog timer. This code does not include the clocksource
nor IRQ features.

The driver does not try to touch anything as long as no timer
allocation is requested. This limits the risk of side effects
in presence of broken BIOSes for instance, which reset the timers
while they shouldn't. In such a case, simply passing "mfgptfix"
to the kernel command line is enough to fix the problem.

When the driver was backported, 2.6 driver was at commit ID
0d5cdc97e242a5589e5dca23277675f4b4482490.

Signed-off-by: Willy Tarreau <w@1wt.eu>