exit_notify: fix regression uncovered by the CAP_KILL fix
Patch 3551c6a671cd7aa17ef6254276000ab8f3171708 silently uncovered
a bug waiting to strike. The problem is that request_module() does
not set the modprobe task's exit_signal and does not expect it to
send a signal when it dies, as it does a waitpid(pid, NULL, __WCLONE).

During exit_notify(), exit signal 0 was not changed when modprobe
exited because modprobe had CAP_KILL set ! Since we removed that
test, modprobe sees its exit_signal assigned to SIGCHLD, which does
not please its parent and leaves modprobe zombies everywhere.

The correct fix consists in not setting exit_signal to SIGCHLD if
exit_signal was not previously set.

This simple reproducer easily triggers the bug, with CONFIG_IPV6=m :

  #include <stdio.h>
  #include <sys/socket.h>
  #include <netinet/in.h>

  main() {
	socket(PF_INET6, SOCK_STREAM, IPPROTO_TCP);
        getchar();
  }

It causes the following line to be displayed in kernel logs :

  request_module[net-pf-10]: waitpid(371,...) failed, errno 1

A modprobe zombie lies around until the reproducer dies.

Special thanks to Michael Niehren for reporting the issue and helping
in quickly narrowing the problem down by testing patches.

Signed-off-by: Willy Tarreau <w@1wt.eu>