r8169: fix erroneous receive packet size settings
Eventhough the r8169 driver in kernel 2.4 is not directly affected
by CVE-2009-1389, it has some bugs which make it sensible to similar
stimuli. More specifically, when receive descriptors are recycled,
their announced size is not correctly set : instead of being set,
the size is ORed with the size of the last received packet. Since
the configured packet size is 1536 (0x600), we can have up to 0x7FF
(0x600|0x5ff) configured in a descriptor after it is recycled,
causing overflows of 511 bytes.

A second issue was that the NET_IP_ALIGN value was not accounted in
the call to alloc_skb().

Note that this fix alone is not enough, as fragmented frames do
still overflow the receive socket.

Check Eric Dumazet's patch for 2.6 for more information on the
subject (2.6 commit fdd7b4c3302c93f6833e338903ea77245eb510b4).

Signed-off-by: Willy Tarreau <w@1wt.eu>