r8169: reject fragmented frames to prevent panics with large frames
This patch follows commit 8d5555548d0b2c57d798efd340b92706912fe138
which fixes crashes when receiving large frames.

When the r8169 receives frames larger than a receive descriptor, it
fragments it across multiple descriptors which are stuffed with 0x3FF0
instead of the real size. This causes the driver to interprete these
fragments as 8 kB packets (larger than the allocated skb), leading to
an immediate kernel panic during the call to skb_put().

The solution consists in rejecting fragmented frames, just as is done
in 2.6.

Signed-off-by: Willy Tarreau <w@1wt.eu>