r8169: reset the chip on receive fifo overflows
When the chip receives larger frames than configured, it spans them
across multiple fragments, causing very quick shortage of receive
descriptors, which translates into RX fifo overflows, after which
the chip cannot receive anything anymore. The only way to get out
of this situation is to reset it. Since this cannot be done in
interrupt context, we have to wake a dedicated task for this.

Some parts of the reset logic were taken from the 2.6 driver which
already does this.

With this patch applied, the driver now correctly recovers from
any unexpected situation caused by large incoming frames.

Signed-off-by: Willy Tarreau <w@1wt.eu>