usb-storage: fix "READ CAPACITY failed" errors with USB flash drives
It's very common that inserting a USB flash device on a machine
running kernel 2.4 causes a "READ CAPACITY failed" error message,
and some devices require multiple attempts to be correctly detected.
Waiting 1 second once the device is detected is often more than
enough to definitely get rid of the problem.

I've been using this patch for over a year with a lot of success,
so let's merge it.

Signed-off-by: Willy Tarreau <w@1wt.eu>