build: fix genksyms segfault in pcigame.c
Solar Designer reported a genksyms segfault. After investigation,
it turned out this was caused by defines overriding function
definitions in pcigame.c when CONFIG_INPUT_PCIGAME is disabled. Since
this only happens during make dep with MODVERSIONS and not when building
the module, the fix simply consists in undefining the offending names
in the C file itself.

Signed-off-by: Willy Tarreau <w@1wt.eu>