restrict reading from /proc/<pid>/maps to those who share ->mm or can ptrace pid
This work was inspired by the following 2.6 commit :

  commit 831830b5a2b5d413407adf380ef62fe17d6fcbf2
  Author: Al Viro <viro@ZenIV.linux.org.uk>
  Date:   Wed Jan 2 14:09:57 2008 +0000

  Contents of /proc/*/maps is sensitive and may become sensitive after
  open() (e.g.  if target originally shares our ->mm and later does exec
  on suid-root binary).

  Check at read() (actually, ->start() of iterator) time that mm_struct
  we'd grabbed and locked is
   - still the ->mm of target
   - equal to reader's ->mm or the target is ptracable by reader.

The fix was adapted not to depend on __ptrace_may_attach(), which does
not exist in 2.4. Since we don't support mmu-less systems here, the fix
was much more straight forward than it was in 2.6.

Note that due to the fact that we still have the task_lock(), we are
not concerned by the race fixed by 2.6 commit 704b836cbf19e885f8366bccb2e4b0474346c02d.

Signed-off-by: Willy Tarreau <w@1wt.eu>